PP-4373 Separate 3DS controller
Separates out 3DS controller from charge controller, as its
a fairly self contained unit. I have not made any other changes
other than lifting and shifting, and amending the necessary routes.
