+++ app/controllers/charge_controller.js
@@ -23,7 +23,7 @@ const connectorClient = require('../services/clients/connector_client')
 
 // constants
 const clsXrayConfig = require('../../config/xray-cls')
-const {views, threeDsEPDQResults, preserveProperties} = require('../../config/charge_controller')
+const {views, preserveProperties} = require('../../config/charge_controller')
 const {CORRELATION_HEADER} = require('../../config/correlation_header')
 
 const appendChargeForNewView = (charge, req, chargeId) => {
@@ -51,26 +51,6 @@ const redirect = res => {
     toNew: (chargeId) => res.redirect(303, routeFor('new', chargeId)),
     toReturn: (chargeId) => res.redirect(303, routeFor('return', chargeId))
   }
-}
-
-const build3dsPayload = req => {
-  let auth3dsPayload = {}
-  const paRes = _.get(req, 'body.PaRes')
-  if (!_.isUndefined(paRes)) {
-    auth3dsPayload.pa_response = paRes
-  }
-
-  const providerStatus = threeDsEPDQResults[_.get(req, 'body.providerStatus', '')]
-  if (!_.isUndefined(providerStatus)) {
-    auth3dsPayload.auth_3ds_result = providerStatus
-  }
-
-  const md = _.get(req, 'body.MD')
-  if (!_.isUndefined(md)) {
-    auth3dsPayload.md = md
-  }
-
-  return auth3dsPayload
 }
 
 const handleCreateResponse = (req, res, charge) => response => {
@@ -128,24 +108,6 @@ const handleAuthResponse = (req, res, charge) => response => {
       break
     default:
       redirect(res).toNew(req.chargeId)
-  }
-}
-
-const handleThreeDsResponse = (req, res, charge) => response => {
-  switch (response.statusCode) {
-    case 200:
-    case 400:
-      redirect(res).toConfirm(charge.id)
-      break
-    case 202:
-    case 409:
-      redirect(res).toAuthWaiting(charge.id)
-      break
-    case 500:
-      responseRouter.response(req, res, 'SYSTEM_ERROR', withAnalytics(charge))
-      break
-    default:
-      responseRouter.response(req, res, 'ERROR', withAnalytics(charge))
   }
 }
 
@@ -314,60 +276,6 @@ module.exports = {
       default:
         redirect(res).toConfirm(req.chargeId)
     }
-  },
-  auth3dsHandler (req, res) {
-    const charge = normalise.charge(req.chargeData, req.chargeId)
-    const correlationId = req.headers[CORRELATION_HEADER] || ''
-    const payload = build3dsPayload(req)
-    connectorClient({correlationId}).threeDs({chargeId: charge.id, payload})
-      .then(handleThreeDsResponse(req, res, charge))
-      .catch(() => {
-        responseRouter.response(req, res, 'ERROR', withAnalytics(charge))
-      })
-  },
-  auth3dsRequired: (req, res) => {
-    const charge = normalise.charge(req.chargeData, req.chargeId)
-    responseRouter.response(req, res, views.AUTH_3DS_REQUIRED_VIEW, withAnalytics(charge))
-  },
-  auth3dsRequiredOut: (req, res) => {
-    const charge = normalise.charge(req.chargeData, req.chargeId)
-    const issuerUrl = _.get(charge, 'auth3dsData.issuerUrl')
-    const paRequest = _.get(charge, 'auth3dsData.paRequest')
-    const md = _.get(charge, 'auth3dsData.md')
-    const htmlOut = _.get(charge, 'auth3dsData.htmlOut')
-
-    if (issuerUrl && paRequest) {
-      let data = {
-        issuerUrl: issuerUrl,
-        paRequest: paRequest,
-        threeDSReturnUrl: `${req.protocol}://${req.hostname}${paths.generateRoute('external.card.auth3dsRequiredIn', {chargeId: charge.id})}`
-      }
-      if (md) {
-        data.md = md
-      }
-      responseRouter.response(req, res, views.AUTH_3DS_REQUIRED_OUT_VIEW, data)
-    } else if (htmlOut) {
-      responseRouter.response(req, res, views.AUTH_3DS_REQUIRED_HTML_OUT_VIEW, {
-        htmlOut: Buffer.from(htmlOut, 'base64').toString('utf8')
-      })
-    } else {
-      responseRouter.response(req, res, 'ERROR', withAnalytics(charge))
-    }
-  },
-  auth3dsRequiredIn: (req, res) => {
-    const charge = normalise.charge(req.chargeData, req.chargeId)
-    responseRouter.response(req, res, views.AUTH_3DS_REQUIRED_IN_VIEW, {
-      threeDsHandlerUrl: routeFor('auth3dsHandler', charge.id),
-      paResponse: _.get(req, 'body.PaRes'),
-      md: _.get(req, 'body.MD')
-    })
-  },
-  auth3dsRequiredInEpdq: (req, res) => {
-    const charge = normalise.charge(req.chargeData, req.chargeId)
-    responseRouter.response(req, res, views.AUTH_3DS_REQUIRED_IN_VIEW, {
-      threeDsHandlerUrl: routeFor('auth3dsHandler', charge.id),
-      providerStatus: _.get(req, 'query.status', 'success')
-    })
   },
   confirm: (req, res) => {
     const charge = normalise.charge(req.chargeData, req.chargeId)
